Cleanup of sending directory, small fixes
Renamed the directories to be more consistent with the receiver, and removed all instances of the LED code for now, since it was causing problems with sending packets. Will integrate it back in later

To be more consistent with the whole naming structure

Cleanup sending directory structure, removed LED

Was interfering with normal functions, will go back and integrate it later.