Refactored website/ dir, removed unused code
Particularly, there was a lot of old JavaScript that was used to parse incoming raw bytes, because it's really painful to do that in JavaScript. Fortunately, we found a much better way of doing it, so we just cut out all the old functions that dealt with the weird & hacky way.