Fix bandwidth not 0-ing, unresponsive web buttons
The bandwidth calculation on the receiving side would not decrease, due to the way that we were calculating the bandwidth (i.e. only when we received a packet). The bandwidth calculation is now moved to whenever we request it.

Hopefully, that one button that would not respond is now fixed. I'm pretty sure it was due to the col that it was in not being wide enough, so it wouldn't register the mouseover or the mouseclick on parts of the button not inside the designated area.